requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
atlassian-python-api>=3.41.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
click>=8.0.0
anthropic>=0.25.0
google-generativeai>=0.3.0

# Development dependencies
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
